fix: resolve signed/unsigned comparison warnings in OtlpRecordable

+++ exporters/otlp/src/otlp_recordable.cc
@@ -120,7 +120,7 @@ void OtlpRecordable::SetResource(const sdk::resource::Resource &resource) noexce
 void OtlpRecordable::SetAttribute(nostd::string_view key,
                                   const common::AttributeValue &value) noexcept
 {
-  if (span_.attributes_size() >= max_attributes_)
+  if (static_cast<uint32_t>(span_.attributes_size()) >= max_attributes_)
   {
     span_.set_dropped_attributes_count(span_.dropped_attributes_count() + 1);
     return;
@@ -134,7 +134,7 @@ void OtlpRecordable::AddEvent(nostd::string_view name,
                               common::SystemTimestamp timestamp,
                               const common::KeyValueIterable &attributes) noexcept
 {
-  if (span_.events_size() >= max_events_)
+  if (static_cast<uint32_t>(span_.events_size()) >= max_events_)
   {
     span_.set_dropped_events_count(span_.dropped_events_count() + 1);
     return;
@@ -145,7 +145,7 @@ void OtlpRecordable::AddEvent(nostd::string_view name,
   event->set_time_unix_nano(timestamp.time_since_epoch().count());
 
   attributes.ForEachKeyValue([&](nostd::string_view key, common::AttributeValue value) noexcept {
-    if (event->attributes_size() >= max_attributes_per_event_)
+    if (static_cast<uint32_t>(event->attributes_size()) >= max_attributes_per_event_)
     {
       event->set_dropped_attributes_count(event->dropped_attributes_count() + 1);
       return true;
@@ -158,7 +158,7 @@ void OtlpRecordable::AddEvent(nostd::string_view name,
 void OtlpRecordable::AddLink(const trace::SpanContext &span_context,
                              const common::KeyValueIterable &attributes) noexcept
 {
-  if (span_.links_size() >= max_links_)
+  if (static_cast<uint32_t>(span_.links_size()) >= max_links_)
   {
     span_.set_dropped_links_count(span_.dropped_links_count() + 1);
     return;
@@ -170,7 +170,7 @@ void OtlpRecordable::AddLink(const trace::SpanContext &span_context,
                     trace::SpanId::kSize);
   link->set_trace_state(span_context.trace_state()->ToHeader());
   attributes.ForEachKeyValue([&](nostd::string_view key, common::AttributeValue value) noexcept {
-    if (link->attributes_size() >= max_attributes_per_link_)
+    if (static_cast<uint32_t>(link->attributes_size()) >= max_attributes_per_link_)
     {
       link->set_dropped_attributes_count(link->dropped_attributes_count() + 1);
       return true;
